% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode-package.R
\docType{package}
\name{geocode-package}
\alias{geocode-package}
\title{geocode: A package for computating the notorious bar statistic.}
\description{
The geocode package provides two categories of important functions:
\code{\link{geocode}}, and \code{\link{clean_addr}}.
}
