% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\alias{get_gps_}
\title{Get GPS (lon, lat) from address}
\usage{
geocode(addrs, source = "google", n_cpu = -1L, rate = 200,
  use_tor = TRUE)

get_gps_(addr, rate = 200, use_tor = FALSE, ...)
}
\arguments{
\item{rate}{Connection rate per second}

\item{addr}{Address string}
}
\value{
GPS data.table
}
\description{
Get GPS (lon, lat) from address
}
\section{Functions}{
\itemize{
\item \code{geocode}: Get GPS from address vector

\item \code{get_gps_}: Get GPS from address vector length of one
}}

\examples{
addrs <- c("台北市中正區羅斯福路一段２號",
           "台北市中正區貴陽街一段１２０號")
geocode(addrs)
}
