% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geocode.R, R/geocode.R
\name{add_geocode}
\alias{add_geocode}
\alias{geocode}
\title{Get GPS (lon, lat) from address}
\usage{
add_geocode(data, addr_var, precise = FALSE, source = "google",
  parallel = TRUE, n_cpu = -1L, rate = 200, use_tor = TRUE)

geocode(addr, precise = FALSE, source = "google", parallel = TRUE,
  n_cpu = -1L, rate = 200, use_tor = TRUE)
}
\arguments{
\item{data}{data.frame contains address variable.}

\item{addr_var}{Character. Name of address variable.}

\item{source}{API source}

\item{parallel}{Parallel connection}

\item{n_cpu}{}

\item{rate}{Connection rate per second}

\item{use_tor}{Boolean. Default TRUE. Use tor or not.}

\item{addr}{Address string}
}
\value{
GPS data.table
}
\description{
Get GPS (lon, lat) from address
}
\section{Functions}{
\itemize{
\item \code{add_geocode}: Add GIS columns to data.frame.

\item \code{geocode}: Get GPS from address vector
}}

\examples{
library(data.table)
library(magrittr)

dt <- data.table(addrs = c("台北市中正區羅斯福路一段２號",
                           "台北市中正區貴陽街一段１２０號"),
                 X = c(3, 2))
out <- dt \%>\% add_geocode(addr_var = "addrs", use_tor = FALSE)
out

addr <- c("台北市中正區羅斯福路一段２號",
           "台北市中正區貴陽街一段１２０號")
geocode(addr)
}
